fix(intake): repair pylint and pyright seams
Why:
- pylint was missing or flagging files in the intake and skill-script seams, and pyright was surfacing cross-module private-usage and import-cycle errors.

What:
- move the capture-session models into intake planning, publish the Ronin translation helpers across the package boundary, update the pylint target coverage contract, and make the skill scripts bootstrap their imports without lint violations.

Checks:
- UV_PROJECT_ENVIRONMENT="$HOME/.venvs/tallylot-py312" uv run python -m tools.run_quality_gates --full-tests
