+++ src/tallylot/application/intake/plan/models.py
@@ -1,13 +1,12 @@
-"""Typed rows and headers for intake plan artifacts."""
+"""Typed rows, headers, and capture-session models for intake plan artifacts."""
 
 from __future__ import annotations
 
+from collections.abc import Sequence
 from dataclasses import dataclass
 from pathlib import Path
-from typing import TYPE_CHECKING
 
-if TYPE_CHECKING:
-    from tallylot.application.intake.captures.session import CaptureSessionPlan
+from tallylot.domain.types import JsonValue
 
 PLAN_HEADER = (
     "path",
@@ -120,6 +119,83 @@ class PlannedItem:
             "inventory_match_status": self.inventory_match_status,
             "target_path": str(self.target_path),
         }
+
+
+@dataclass(frozen=True)
+class CaptureSessionPlan:
+    source_folder: str
+    capture_label: str
+    manifest_fingerprint: str
+    capture_status: str
+    file_count: int
+    observed_period_start: str
+    observed_period_end: str
+    observed_group_count: int
+    duplicate_capture_uid: str = ""
+    overlap_capture_uids: tuple[str, ...] = ()
+
+    def to_summary(
+        self,
+        *,
+        planned_items: Sequence[PlannedItem],
+        issue_rows: list[dict[str, str]],
+    ) -> dict[str, JsonValue]:
+        return {
+            "source": self.source_folder,
+            "planned_capture_label": self.capture_label,
+            "manifest_fingerprint": self.manifest_fingerprint,
+            "capture_status": self.capture_status,
+            "observed_period_start": self.observed_period_start,
+            "observed_period_end": self.observed_period_end,
+            "observed_group_count": self.observed_group_count,
+            "duplicate_capture_uid": self.duplicate_capture_uid,
+            "overlap_capture_uids": list(self.overlap_capture_uids),
+            "file_count": self.file_count,
+            "issue_count": len(issue_rows),
+            "planned_copy_count": sum(
+                1 for item in planned_items if item.action in {"copy", "extract_copy"}
+            ),
+            "copied_count": 0,
+            "explicit_map_count": sum(
+                1
+                for item in planned_items
+                if item.source_resolution_status == "explicit_map"
+            ),
+            "explicit_map_blocked_count": sum(
+                1
+                for item in planned_items
+                if item.source_resolution_status == "explicit_map_blocked"
+            ),
+            "source_label_map_issue_count": sum(
+                1 for row in issue_rows if row["kind"].startswith("source_label_map_")
+            ),
+            "duplicate_packages": _package_count(planned_items, "duplicate_package"),
+            "merge_primary_packages": _package_count(planned_items, "merge_primary"),
+            "merged_packages": _package_count(planned_items, "merge_member"),
+            "overlap_packages": _package_count(planned_items, "overlap_partial_review"),
+            "mixed_cycle_packages": _package_count(planned_items, "mixed_cycle_review"),
+        }
+
+
+@dataclass(frozen=True)
+class CaptureSessionSummaryContext:
+    planned_items: Sequence[PlannedItem]
+    issue_rows: list[dict[str, str]]
+    copied_count: int
+
+
+def _package_count(planned_items: Sequence[PlannedItem], package_status: str) -> int:
+    return len(
+        {
+            (item.source_folder, item.capture_label, item.bundle_id)
+            for item in planned_items
+            if item.package_status == package_status
+            or (
+                package_status == "duplicate_package"
+                and item.package_status.startswith("duplicate_package")
+            )
+        }
+    )
 
 
 @dataclass(frozen=True)
